Add Google Fonts and enhanced theme properties in TailwindCSS config
This commits the addition of Google Font "Hanken Grotesk" to the index.html. This is done by a utilization of a CDN (Content Delivery Network) link, allowing for more flexible and diverse font choices. Concurrently, there are changes in the tailwind.config.js where theme properties have been extended. This extends the available color palette while specifying the newly added font-family which gives more scope for design choice and boosts the design consistency throughout the website.